Delete ChatRoomContainer
chat room container 삭제
