+++ packages/client/src/pages/chatroom/index.tsx
@@ -1,34 +1,76 @@
 import React, { FC } from 'react'
-import { WithAuthProps } from 'hocs/withAuth'
-import { ChatStateGroupByTime } from 'containers/ChatRoomContainer'
+import withAuth, { WithAuthProps } from 'hocs/withAuth'
+import { useRouteMatch } from 'react-router-dom'
 import SearchAccordion from 'system/ChatRoomSearchBar'
-import * as S from './style'
+import { getChatRequest } from 'modules/chat'
+import {
+  useDispatch, useSelector,
+} from 'react-redux'
+import { RootState } from 'modules'
+import {
+  chatFromServer,
+  Event,
+  joinRooms,
+  removeSocketEventListener,
+} from 'socket'
 import Header from './Header'
 import TextArea from './TextArea'
 import ChatArea from './ChatArea'
+import * as S from './style'
 
 const {
-  useEffect, useRef, useState, useCallback,
+  useEffect, useState, useCallback,
 } = React
 
-interface Props extends WithAuthProps{
-  roomUuid: string
-  roomName: string
-  chatStateGroupByTime: ChatStateGroupByTime,
-}
-
-const ChatRoom: FC<Props> = ({
-  uuid,
-  roomUuid,
-  roomName,
-  chatStateGroupByTime,
-}) => {
+const ChatRoom: FC<WithAuthProps> = ({ uuid }) => {
   const [isSearchOpen, setIsSearchOpen] = useState(false)
+  const { params } = useRouteMatch<{roomUuid: string}>()
+  const [roomUuid, setRoomUuid] = useState('')
+  const [roomName, setRoomName] = useState<string>('')
+  const dispatch = useDispatch()
+  const roomState = useSelector((state: RootState) => state.room)
 
   const toggleSearchBar = useCallback(() => {
     setIsSearchOpen(!isSearchOpen)
   }, [isSearchOpen, setIsSearchOpen])
 
+  useEffect(() => {
+    if (roomUuid.length) {
+      joinRooms({ roomUuids: [roomUuid] })
+    }
+  }, [roomUuid])
+
+  useEffect(() => {
+    setRoomUuid(params.roomUuid)
+    const limit = 15
+    const offset = 0
+    dispatch(getChatRequest({
+      roomUuid: params.roomUuid,
+      limit,
+      offset,
+    }))
+  }, [params.roomUuid, dispatch])
+
+  useEffect(() => {
+    const rn = roomState.data.find((v) => v.uuid === roomUuid)
+    if (!rn) {
+      return
+    }
+
+    setRoomName(rn.participants.map((v) => v.name).join(', '))
+  }, [roomState, roomUuid])
+
+  useEffect(() => {
+    setRoomUuid(params.roomUuid)
+  }, [params])
+
+  useEffect(() => {
+    chatFromServer(dispatch)
+
+    return (() => {
+      removeSocketEventListener(Event.chatFromServer)
+    })
+  })
   return (
     <S.Container>
       <Header roomName={roomName} toggleSearchBar={toggleSearchBar}/>
@@ -42,4 +84,4 @@ const ChatRoom: FC<Props> = ({
   )
 }
 
-export default ChatRoom
+export default withAuth(ChatRoom)
